Constraints... seem to work now

+++ src/types.rs
@@ -17,15 +17,15 @@ pub enum Task {
     },
     Baseline {
         name: Name,
-        skills: HashMap<Skill, f64>,
+        skills: HashMap<Skill, f32>,
     },
     Schedule {
         name: Name,
-        segment: HashMap<Segment, f64>,
+        segment: HashMap<Segment, f32>,
     },
     SafetyLimit {
         name: Name,
-        limit: HashMap<Skill, f64>,
+        limit: HashMap<Skill, f32>,
     },
     ScheduleLimit {
         name: Name,
@@ -37,23 +37,30 @@ pub enum Task {
     },
     Target {
         name: Name,
-        target: HashMap<Skill, f64>,
+        target: HashMap<Skill, f32>,
     },
 }
 
 #[derive(Debug)]
 pub struct Person {
     pub name: Name,
-    pub skills: HashMap<Skill, f64>,
-    pub schedule: HashMap<Segment, f64>,
-    pub safety_limit: HashMap<Skill, f64>,
+    // This person's skills, training or not.
+    pub skills: HashMap<Skill, f32>,
+    // This person's schedule, in terms of segments and their duration.
+    pub schedule: HashMap<Segment, f32>,
+    // Limits to how much some skills can be trained per day.
+    pub safety_limit: HashMap<Skill, f32>,
+    // Limits to which skills can be trained in which segments.
+    // Some segments have no limit, and are not listed here.
     pub schedule_limit: HashMap<Segment, Vec<Skill>>,
+    // Overlap bonuses for training multiple skills at once.
+    // This *includes* the trivial case of training a single skill.
     pub overlap: Vec<Overlap>,
-    pub target: HashMap<Skill, f64>,
+    pub target: HashMap<Skill, f32>,
 }
 
 impl Person {
-    pub fn new(name: Name, skills: HashMap<Skill, f64>) -> Self {
+    pub fn new(name: Name, skills: HashMap<Skill, f32>) -> Self {
         Self {
             name,
             skills,
@@ -69,5 +76,5 @@ impl Person {
 #[derive(Debug)]
 pub struct Overlap {
     pub combo: Vec<Skill>,
-    pub bonus: f64,
+    pub bonus: f32,
 }
